BUGFIX: NaNs appearing in nearfield approximation
The nearfield approximation is executed on integration points that match
a given geometrical criterion to be "near" the surface element.
If within this criterion there is no integration point, the
NearFieldCharge as well as the NearFieldChargeArea are set to zero, thus
leading to the appearance of NaNs in the flux container. A simple check
for reasonable values avoids this.